added flag check instead of (.Capabilities.APIVersions.Has "gateway.networking.k8s.io/v1") check

+++ facade-operator-service/controllers/facade_common_controller.go
@@ -6,6 +6,9 @@ import (
 	"runtime/debug"
 
 	"fmt"
+	"strconv"
+	"time"
+
 	"github.com/netcracker/qubership-core-facade-operator/facade-operator-service/v2/api/facade"
 	facadeV1Alpha "github.com/netcracker/qubership-core-facade-operator/facade-operator-service/v2/api/facade/v1alpha"
 	customerrors "github.com/netcracker/qubership-core-facade-operator/facade-operator-service/v2/pkg/errors"
@@ -16,8 +19,6 @@ import (
 	"github.com/netcracker/qubership-core-facade-operator/facade-operator-service/v2/pkg/utils"
 	errs "github.com/netcracker/qubership-core-lib-go-error-handling/v3/errors"
 	"github.com/netcracker/qubership-core-lib-go/v3/logging"
-	"strconv"
-	"time"
 
 	v1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -43,7 +44,6 @@ type FacadeCommonReconciler struct {
 	hpaClient          services.HPAClient
 	ingressClient      services.IngressClientAggregator
 	httpRouteClient    services.HTTPRouteClient
-	gwAPIEnabled       bool
 	ingressBuilder     *templates.IngressTemplateBuilder
 	namedLock          *utils.NamedResourceLock
 	controlPlaneClient restclient.ControlPlaneClient
@@ -68,7 +68,6 @@ func NewFacadeCommonReconciler(
 	readyService services.ReadyService,
 	commonCRClient services.CommonCRClient,
 	crPriorityService services.CRPriorityService,
-	gatewayAPIV1Present bool,
 ) *FacadeCommonReconciler {
 	return &FacadeCommonReconciler{
 		client:             client,
@@ -80,7 +79,6 @@ func NewFacadeCommonReconciler(
 		hpaClient:          hpaClient,
 		ingressClient:      ingressClient,
 		httpRouteClient:    services.NewHTTPRouteClient(client, ingressBuilder, commonCRClient),
-		gwAPIEnabled:       gatewayAPIV1Present,
 		ingressBuilder:     ingressBuilder,
 		controlPlaneClient: controlPlaneClient,
 		namedLock:          utils.NewNamedResourceLock(),
@@ -180,10 +178,12 @@ func (r *FacadeCommonReconciler) deleteFacadeService(ctx context.Context, req ct
 		return err
 	}
 
-	if err := r.ingressClient.DeleteOrphaned(ctx, req); err != nil {
-		return err
+	if utils.ShouldDeployIngress() {
+		if err := r.ingressClient.DeleteOrphaned(ctx, req); err != nil {
+			return err
+		}
 	}
-	if r.gwAPIEnabled {
+	if utils.ShouldDeployGatewayAPI() {
 		if err := r.httpRouteClient.DeleteOrphaned(ctx, req); err != nil {
 			return err
 		}
@@ -325,10 +325,12 @@ func (r *FacadeCommonReconciler) applyFacadeService(ctx context.Context, req ctr
 }
 
 func (r *FacadeCommonReconciler) applyIngresses(ctx context.Context, req ctrl.Request, serviceName string, cr facade.MeshGateway) error {
-	if err := r.ingressClient.DeleteOrphaned(ctx, req); err != nil {
-		return err
+	if utils.ShouldDeployIngress() {
+		if err := r.ingressClient.DeleteOrphaned(ctx, req); err != nil {
+			return err
+		}
 	}
-	if r.gwAPIEnabled {
+	if utils.ShouldDeployGatewayAPI() {
 		if err := r.httpRouteClient.DeleteOrphaned(ctx, req); err != nil {
 			return err
 		}
@@ -341,15 +343,17 @@ func (r *FacadeCommonReconciler) applyIngresses(ctx context.Context, req ctrl.Re
 
 func (r *FacadeCommonReconciler) applyIngressSpecs(ctx context.Context, req ctrl.Request, serviceName string, cr facade.MeshGateway) error {
 	for _, ingressSpec := range cr.GetSpec().Ingresses {
-		ingressTemplate, err := r.ingressBuilder.BuildIngressTemplate(ingressSpec, cr, serviceName)
-		if err != nil {
-			return err
+		if utils.ShouldDeployIngress() {
+			ingressTemplate, err := r.ingressBuilder.BuildIngressTemplate(ingressSpec, cr, serviceName)
+			if err != nil {
+				return err
+			}
+			r.logger.InfoC(ctx, "[%v] Applying ingress %+v", req.NamespacedName, ingressSpec)
+			if err = r.ingressClient.Apply(ctx, req, ingressTemplate); err != nil {
+				return err
+			}
 		}
-		r.logger.InfoC(ctx, "[%v] Applying ingress %+v", req.NamespacedName, ingressSpec)
-		if err = r.ingressClient.Apply(ctx, req, ingressTemplate); err != nil {
-			return err
-		}
-		if err = r.applyHTTPRouteIfEnabled(ctx, req, serviceName, cr, ingressSpec); err != nil {
+		if err := r.applyHTTPRouteIfEnabled(ctx, req, serviceName, cr, ingressSpec); err != nil {
 			return err
 		}
 	}
@@ -357,7 +361,7 @@ func (r *FacadeCommonReconciler) applyIngressSpecs(ctx context.Context, req ctrl
 }
 
 func (r *FacadeCommonReconciler) applyHTTPRouteIfEnabled(ctx context.Context, req ctrl.Request, serviceName string, cr facade.MeshGateway, ingressSpec facade.IngressSpec) error {
-	if r.gwAPIEnabled {
+	if utils.ShouldDeployGatewayAPI() {
 		httpRouteTemplate, err := r.ingressBuilder.BuildHTTPRouteTemplate(ingressSpec, cr, serviceName)
 		if err != nil {
 			return err
